updated back end
added authRoute
