+++ server/routes/userRoute.js
@@ -4,7 +4,7 @@ const authMiddleware=require('../middleware/authMiddleware')
 const userController = require('../controllers/userController');
 
 // User routes
-router.post('/login', userController.login);            // GET /users hello  hee- Get all users
+          // GET /users hello  hee- Get all users
 router.post('/createnewuser',userController.createNewUser)
-router.post('/getUserData',authMiddleware,userController.authController)
+
 module.exports = router;
